Implement proper fence ring enclosure for horse containment in tests
- Build a 5x5 fence perimeter at ground level with open 3x3 center
- Fence is constructed before horse summon and removed after mount
- Prevents horse wandering while allowing normal mounting interaction
- No interference with vertical movement or entity physics

+++ items/usage_test.go
@@ -358,7 +358,7 @@ func TestAttackEntity(t *testing.T) {
 func TestSwitchToSlot(t *testing.T) {
 	tests := []struct {
 		name        string
-		slotIndex   int
+		slotIndex   int16
 		shouldSend  bool
 		description string
 	}{
